pyproject.toml
[project]
name = "eclipce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.16",
    "aiomysql>=0.2.0",
    "bcrypt>=4.3.0",
    "cryptography>=44.0.2",
    "fastapi[all]>=0.115.12",
    "pycryptodome>=3.22.0",
    "pyjwt>=2.10.1",
    "python-gnupg>=0.5.4",
    "redis[hiredis]>=5.2.1",
    "requests>=2.32.3",
    "setuptools>=78.1.1",
    "sqlalchemy>=2.0.40",
]
